requirements.txt
numpy==1.21.6
pandas==1.3.5
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2026.1.post1
PyYAML==6.0.1
six==1.17.0
